debug statements, move another fn into platform mod
